+++ script/test_cljfmt.py
@@ -9,10 +9,10 @@ cs_cljfmt = importlib.import_module(module + '.cs_cljfmt')
 sublime = importlib.import_module(module + '.script.sublime')
 test_core = importlib.import_module(module + '.script.test_core')
 
-def test_printer():
+def test():
     def test_fn(input):
         return cs_cljfmt.format_string(input, cwd = dir)
     test_core.run_tests(dir + "/test_indent/", test_fn, col_input = False)
 
 if __name__ == '__main__':
-    test_printer()
+    test()